Add jujutsu techniques to data
It's not shown yet because I haven't decided on how to split or not
split judo and jujutsu.

Also includes longer "instructional" videos. I might add some way to
have the short video as a reminder for the technique assuming the user
knows the technique already, but somehow show the longer video if one
wants to learn the technique.