Move sort logic out of the reducers
Sorting and filtering of techniques is now a helper method, and used by
TechniquesTable to get the correct content and sorting. The random
technique reducer should also be refactored in a similar fashion.

CategoryCheckboxes now uses dataFriendly values and the default
selection is all of them checked (instead of null).