Restructure techniques in store
Handling random cards is easier if it's just a random sort order instead
of "the next random card". That allows the app to potentially have a
"previous card" action and a progress meter. As well as keeping logic
out of actions, but at the cost of complicating the techniques reducer.